Add road construction sites where creeps commonly travel
This implementation is incredibly CPU heavy (as I loop through possible
positions in a really dumb way) - as such, the calculations only done every 300
ticks.  It will still often overflow on that tick, causing no new position to
be calculated.

A config or programmatic way of setting calculation frequency would be good as
the program gets more complex.

Also, the creep position object should eventually not be keyed by the
RoomPosition object, and should be wiped often (or perhaps, rolling weighted
values?)